# Conflicts:
#	src/test/java/org/ssu/edu/teachua/api/club/LeadClubTest.java
#	src/test/java/org/ssu/edu/teachua/utils/providers/DataProviderClub.java

+++ src/main/java/org/ssu/edu/teachua/api/models/task/TaskPutRequest.java
@@ -3,6 +3,8 @@ package org.ssu.edu.teachua.api.models.task;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 public class TaskPutRequest {
@@ -11,7 +13,7 @@ public class TaskPutRequest {
     private String description;
     private String picture;
     private String startDate;
-    private int challengeId;
+    private BigInteger challengeId;
 }
