+++ src/main/java/org/ssu/edu/teachua/api/models/task/TaskResponse.java
@@ -2,14 +2,16 @@ package org.ssu.edu.teachua.api.models.task;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class TaskResponse {
-    private int id;
+    private BigInteger id;
     private String name;
     private String headerText;
     private String description;
     private String picture;
     private String startDate;
-    private int challengeId;
+    private BigInteger challengeId;
 }
